Rename `value::Commitment` to `balance::Commitment`
