+++ crypto/src/symmetric.rs
@@ -6,9 +6,9 @@ use chacha20poly1305::{
 use rand::{CryptoRng, RngCore};
 
 use crate::{
-    ka,
+    balance, ka,
     keys::{IncomingViewingKey, OutgoingViewingKey},
-    note, value,
+    note,
 };
 
 pub const PAYLOAD_KEY_LEN_BYTES: usize = 32;
@@ -114,7 +114,7 @@ impl OutgoingCipherKey {
     /// Use Blake2b-256 to derive an encryption key `ock` from the OVK and public fields.
     pub(crate) fn derive(
         ovk: &OutgoingViewingKey,
-        cv: value::Commitment,
+        cv: balance::Commitment,
         cm: note::Commitment,
         epk: &ka::Public,
     ) -> Self {
