+++ transaction/src/action/spend.rs
@@ -3,9 +3,10 @@ use std::convert::{TryFrom, TryInto};
 use anyhow::Error;
 use bytes::Bytes;
 use penumbra_crypto::{
+    balance,
     proofs::transparent::SpendProof,
     rdsa::{Signature, SpendAuth, VerificationKey},
-    value, Nullifier,
+    Nullifier,
 };
 use penumbra_proto::{transaction, Protobuf};
 
@@ -56,7 +57,7 @@ impl TryFrom<transaction::Spend> for Spend {
 
 #[derive(Clone, Debug)]
 pub struct Body {
-    pub value_commitment: value::Commitment,
+    pub balance_commitment: balance::Commitment,
     pub nullifier: Nullifier,
     pub rk: VerificationKey<SpendAuth>,
 }
@@ -68,7 +69,7 @@ impl From<Body> for transaction::SpendBody {
         let nullifier_bytes: [u8; 32] = msg.nullifier.into();
         let rk_bytes: [u8; 32] = msg.rk.into();
         transaction::SpendBody {
-            value_commitment: Some(msg.value_commitment.into()),
+            value_commitment: Some(msg.balance_commitment.into()),
             nullifier: Bytes::copy_from_slice(&nullifier_bytes),
             rk: Bytes::copy_from_slice(&rk_bytes),
         }
@@ -79,7 +80,7 @@ impl TryFrom<transaction::SpendBody> for Body {
     type Error = Error;
 
     fn try_from(proto: transaction::SpendBody) -> anyhow::Result<Self, Self::Error> {
-        let value_commitment: value::Commitment = proto
+        let value_commitment: balance::Commitment = proto
             .value_commitment
             .ok_or_else(|| anyhow::anyhow!("missing value commitment"))?
             .try_into()?;
@@ -96,7 +97,7 @@ impl TryFrom<transaction::SpendBody> for Body {
             .map_err(|_| anyhow::anyhow!("spend body malformed"))?;
 
         Ok(Body {
-            value_commitment,
+            balance_commitment: value_commitment,
             nullifier,
             rk,
         })
