+++ transaction/src/action/swap_claim.rs
@@ -1,8 +1,8 @@
 use ark_ff::Zero;
 use decaf377::Fr;
+use penumbra_crypto::balance;
 use penumbra_crypto::dex::BatchSwapOutputData;
 use penumbra_crypto::transaction::Fee;
-use penumbra_crypto::value;
 use penumbra_crypto::Nullifier;
 use penumbra_crypto::{proofs::transparent::SwapClaimProof, NotePayload};
 use penumbra_proto::{dex as pb, Protobuf};
@@ -18,7 +18,7 @@ pub struct SwapClaim {
 impl SwapClaim {
     /// Compute a commitment to the value contributed to a transaction by this swap claim.
     /// Will add (f,fee_token) representing the pre-paid fee
-    pub fn value_commitment(&self) -> value::Commitment {
+    pub fn value_commitment(&self) -> balance::Commitment {
         self.body.fee.commit(Fr::zero())
     }
 }
